# packages/ms-swagger/api.yaml
openapi: 3.0.0
info:
  title: Business API
  description: API description
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: local Server
    #Tempory
  - url: http://localhost:3000/api
    description: second Server
  - url: http://localhost:3000/api/v1/business/
    description: local Server
    #Tempory
  - url: http://example.com/
    description: original server
components:
  schemas:
...
    Category:
      type: object
      properties:
        shopId:
          type: string
          example: shop123
        categoryCode:
          type: string
          example: VEGI
        name:
          type: string
          example: vegetables
        description:
          type: string
          example: all edible plant including flowers, fruits, stems, leaves, roots, and seeds
        imageURLs:
          type: array
          items:
            type: string
          example: https://cdn.britannica.com/17/196817-050-6A15DAC3/vegetables.jpg
        status:
          type: string
          example: Active
        createdAt:
          type: date
          example: 2022-01-20
        updatedAt:
          type: date
          example: 2022-01-20
        isPublished:
          type: boolean
          example: false
        lastItemId:
          type: string
          example: 0000
      required:
        - categoryCode
        - name
...
  requestBodies:
...
    CategoryBody:
      description: A JSON object containing Category information
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Category"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /business/Category:
    get:
      deprecated: true
      summary: Returns a list of all Category
      tags:
        - Category
      description: Returns All availble Categorys in the database. use only for the development
      produces:
        - application/json
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: User Error
        404:
          description: Not found error

    post:
      tags:
        - Category
      summary: Create a Category
      description: Create a Category with Category details
      produces:
        - application/json
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/CategoryBody"
      responses:
        201:
          description: Successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: Error
        404:
          description: Not found error
...
  /business/Category/shop:
    get:
      tags:
        - Category
      summary: Get Categories by shopId
      description: Get Category data by the shopId of the current user
      produces:
        - application/json
      security:
        - bearerAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: User Error
        404:
          description: Not found error
  /business/Category/shop/{shopId}:
    get:
      tags:
        - Category
      summary: Search Categories by a shopId
      description: Get Category data of a business by passing the shopId as a parameter
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: shopId
          schema:
            type: string
          required: true
          description: shopId to apply the filter
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: User Error
        404:
          description: Not found error
  /business/Category/name/{name}:
    get:
      tags:
        - Category
      summary: Check Categories by Category name
      description: Check Category by passing the category name whether it's available or not in current user
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
          description: Name of the Category which needs to be checked
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: Category available
        400:
          description: User Error
        404:
          description: Not found error
  /business/Category/category-code-check/{categoryCode}:
    get:
      tags:
        - Category
      summary: Check Categories by categoryCode
      description: Check Category by passing the categoryCode whether it's available or not in current user
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: categoryCode
          schema:
            type: string
          required: true
          description: categoryCode which needs to be checked
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: Category available
        400:
          description: User Error
        404:
          description: Not found error
  /business/Category/get-last-item-id/{catId}:
    get:
      tags:
        - Category
      summary: Get last item of a Category
      description: Get the last item of a Category by passing the CategoryId as a parameter
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: catId
          schema:
            type: string
          required: true
          description: Category Id which needs to be checked
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: item0210
        400:
          description: User Error
        404:
          description: Not found error
  /business/Category/{id}:
    get:
      tags:
        - Category
      summary: Search Categories by Id
      description: Get Category data by passing the Category id as a parameter
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id of the Category
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: User Error
        404:
          description: Not found error

    put:
      tags:
        - Category
      summary: Update Category
      description: Update an existing Category data by passing the Category id as a parameter
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Id of the Category
      requestBody:
        $ref: "#/components/requestBodies/CategoryBody"
      responses:
        201:
          description: Successfully Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: User Error
        404:
          description: Not found error

    delete:
      tags:
        - Category
      summary: Delete Category
      description: Delete an existing Category by passing the Category id as a parameter
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Category Id
      responses:
        201:
          description: Successfully Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: User Error
        404:
          description: Not found error
  /business/Category/delete-with-status/{id}:
    delete:
      tags:
        - Category
      summary: Update status as deleted
      description: delete method to update status as deleted by passing the id as a parameter
      produces:
        - application/json
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: id of the Category
      responses:
        200:
          description: Status Updated as Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        400:
          description: User Error
        404:
          description: Not found error

  #-------------------------------  Customer  -------------------------------
...